chore(core,repo,scripts,tests): update core,repo,scripts,tests
- Changes:
  - .taskmaster/tasks/tasks.json
  - .taskmaster/tasks/tasks_back.json
  - .taskmaster/tasks/tasks_gameplay.json
  - Game.Core.Tests/Domain/SanguoContractInstantiationTests.cs
  - Game.Core.Tests/Domain/ValueObjects/CircularMapPositionTests.cs
  - Game.Core.Tests/Engine/GameEngineCoreConstructorTests.cs
  - Game.Core.Tests/Services/InventoryServiceTests.cs
  - Game.Core/Domain/ValueObjects/CircularMapPosition.cs
  - scripts/python/check_test_naming.py
  - scripts/python/run_dotnet.py
  - scripts/python/update_task2_container.py
  - scripts/python/verify_task_mapping.py
  - (+6 more)

- Rationale:
  - Decouple Task 2 ring index/path calculation from `City` to keep the dependency chain clean (City lands in Task 3).
  - Make circular movement arithmetic overflow-safe for extreme step sizes and lock behavior via unit tests.
  - Keep the quality gate reliable by selecting coverage artifacts from the current test run (avoid stale logs) and by adding focused tests to meet the coverage threshold.
  - Provide a repeatable sc workflow (analyze/build tdd/test/git) with auditable outputs under logs/ci for Taskmaster-driven iterations.

Task: #2.1
ADRs: ADR-0005, ADR-0015, ADR-0018, ADR-0020, ADR-0021, ADR-0024, ADR-0025
Overlay: docs/architecture/overlays/PRD-SANGUO-T2/08/08-功能纵切-T2-三国大富翁闭环.md

+++ Game.Core/Domain/ValueObjects/CircularMapPosition.cs
@@ -42,11 +42,11 @@ public readonly record struct CircularMapPosition
     /// <returns>New position.</returns>
     public CircularMapPosition Advance(int steps)
     {
-        var next = (Current + steps) % TotalPositions;
+        var next = ((long)Current + steps) % TotalPositions;
         if (next < 0)
             next += TotalPositions;
 
-        return new CircularMapPosition(next, TotalPositions);
+        return new CircularMapPosition((int)next, TotalPositions);
     }
 }
